+++ server/db/FUN_PRO_VIEWS/01_get_sections_fields_helper.sql
@@ -3,6 +3,8 @@ DROP FUNCTION IF EXISTS get_sections_fields_helper //
 
 CREATE FUNCTION get_sections_fields_helper(section_id INT, language_id INT, gender_id INT) RETURNS TEXT
 -- section_id -1 returns all sections
+READS SQL DATA
+DETERMINISTIC
 BEGIN 
 	SET @@group_concat_max_len = 32000000;
 	SET @sql = NULL;