+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v6.13.0 (Unpublished)
 ### Bugfix 
  - #482 - fix `CMS Preferences` typo
+ - #480 - `READS SQL DATA DETERMINISTIC` to the functions and procedures
 
 # v6.12.2
 ### New features